JMController integrated into RRServer
