+++ twister2/common/src/java/BUILD
@@ -5,6 +5,7 @@ load("//tools:maven.bzl", "t2_java_lib")
 common_deps_files = [
     "@org_yaml_snakeyaml//jar",
     "@com_google_protobuf_protobuf_java//jar",
+    "@com_google_guava_guava//jar",
     "//twister2/proto:proto-java",
 ]
