+++ twister2/common/src/java/edu/iu/dsc/tws/common/net/tcp/request/RRServer.java
@@ -14,13 +14,12 @@ package edu.iu.dsc.tws.common.net.tcp.request;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.SocketChannel;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import com.google.common.collect.HashBiMap;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.Message;
 
@@ -32,7 +31,24 @@ import edu.iu.dsc.tws.common.net.tcp.StatusCode;
 import edu.iu.dsc.tws.common.net.tcp.TCPMessage;
 
 /**
- * Request response server.
+ * RRServer class is used by Job Master
+ * It works in request/response messaging
+ *
+ * Workers and the client always send a request message, and
+ * JobMaster sends a single response message to each request message
+ *
+ * However, sometimes job master may send messages to workers that are not response messages
+ * For example, a client request message may result in Job Master sending a message to all workers
+ *
+ * Message Format:
+ * RequestID (32 bytes), message type length, message type data, senderID (4 bytes), message data
+ * message type is the class name of the protocol buffer for that message
+ *
+ * RequestID is generated in request senders (worker or client),
+ * and the same requestID is used in the response message.
+ *
+ * When job master sends a message that is not a response to a request,
+ * it uses the DUMMY_REQUEST_ID as the requestID.
  */
 public class RRServer {
   private static final Logger LOG = Logger.getLogger(RRServer.class.getName());
@@ -40,9 +56,14 @@ public class RRServer {
   private Server server;
 
   /**
-   * Client id to channel
+   * worker channels with workerIDs
    */
-  private List<SocketChannel> socketChannels = new ArrayList<>();
+  private HashBiMap<SocketChannel, Integer> workerChannels = HashBiMap.create();
+
+  /**
+   * the job submitting client channel,
+   */
+  private SocketChannel clientChannel;
 
   /**
    * Keep track of the request handler using protocol buffer message types
@@ -60,14 +81,19 @@ public class RRServer {
   private Map<RequestID, SocketChannel> requestChannels = new HashMap<>();
 
   /**
-   * Keep track of request id to worker
+   * Job Master ID
    */
-  private Map<RequestID, Integer> requestToWorkers = new HashMap<>();
+  private int serverID;
 
   /**
-   * The worker id
+   * The client id
    */
-  private int workerId;
+  public static final int CLIENT_ID = -100;
+
+  /**
+   * whether JobMaster assigns workerIDs
+   */
+  private boolean jobMasterAssignsWorkerID;
 
   /**
    * Connection handler
@@ -76,11 +102,12 @@ public class RRServer {
 
   private Progress loop;
 
-  public RRServer(Config cfg, String host, int port, Progress looper,
-                  int wId, ConnectHandler cHandler) {
+  public RRServer(Config cfg, String host, int port, Progress looper, int serverID,
+                  ConnectHandler cHandler, boolean jobMasterAssignsWorkerID) {
     this.connectHandler = cHandler;
-    this.workerId = wId;
     this.loop = looper;
+    this.serverID = serverID;
+    this.jobMasterAssignsWorkerID = jobMasterAssignsWorkerID;
     server = new Server(cfg, host, port, loop, new Handler(), false);
   }
 
@@ -103,58 +130,97 @@ public class RRServer {
 
   /**
    * Send a response to a request id
-   * @param id request id
+   * @param requestID request id
    * @param message message
    * @return true if response was accepted
    */
-  public boolean sendResponse(RequestID id, Message message) {
-    LOG.log(Level.FINEST, String.format("Using channel %s", new String(id.getId())));
+  public boolean sendResponse(RequestID requestID, Message message) {
 
-    if (!requestChannels.containsKey(id)) {
-      LOG.log(Level.SEVERE, "Trying to send response to non-existing request");
+    if (!requestChannels.containsKey(requestID)) {
+      LOG.log(Level.SEVERE, "Trying to send a response to non-existing request");
       return false;
     }
 
-    SocketChannel channel = requestChannels.get(id);
+    SocketChannel channel = requestChannels.get(requestID);
 
     if (channel == null) {
       LOG.log(Level.SEVERE, "Channel is NULL for response");
     }
 
-    if (!socketChannels.contains(channel)) {
+    if (!workerChannels.containsKey(channel)) {
       LOG.log(Level.WARNING, "Failed to send response on disconnected socket");
       return false;
     }
 
-    byte[] data = message.toByteArray();
-    String messageType = message.getDescriptorForType().getFullName();
+    TCPMessage tcpMessage = sendMessage(message, requestID, channel);
 
-    // lets serialize the message
-    int capacity = id.getId().length + data.length + 4 + messageType.getBytes().length + 4;
-    ByteBuffer buffer = ByteBuffer.allocate(capacity);
-    // we send message id, worker id and data
-    buffer.put(id.getId());
-    // pack the name of the message
-    ByteUtils.packString(messageType, buffer);
-    // pack the worker id
-    buffer.putInt(workerId);
-    // pack data
-    buffer.put(data);
-
-    TCPMessage request = server.send(channel, buffer, capacity, 0);
-
-    if (request != null) {
-      requestChannels.remove(id);
+    if (tcpMessage != null) {
+      requestChannels.remove(requestID);
       return true;
     } else {
       return false;
     }
   }
 
+  /**
+   * Send a non-response message to a worker or to the client
+   * @param message message
+   * @return true if response was accepted
+   */
+  public boolean sendMessage(Message message, int targetID) {
+
+    SocketChannel channel;
+    if (targetID == CLIENT_ID) {
+      if (clientChannel == null) {
+        LOG.severe("Trying to send a message to the client, but it has not connected yet.");
+        return false;
+      }
+      channel = clientChannel;
+    } else if (workerChannels.containsValue(targetID)) {
+      channel = workerChannels.inverse().get(targetID);
+    } else {
+      LOG.severe("Trying to send a message to a worker that has not connected yet. workerID: "
+          + targetID);
+      return false;
+    }
+
+    // this is most likely not needed, but just to make sure
+    if (channel == null) {
+      LOG.log(Level.SEVERE, "Channel is NULL for response");
+      return false;
+    }
+
+    // since this is not a request/response message, we put the dummy request id
+    RequestID dummyRequestID = RequestID.DUMMY_REQUEST_ID;
+
+    TCPMessage tcpMessage = sendMessage(message, dummyRequestID, channel);
+
+    return tcpMessage == null ? false : true;
+  }
+
+  private TCPMessage sendMessage(Message message, RequestID requestID, SocketChannel channel) {
+    byte[] data = message.toByteArray();
+    String messageType = message.getDescriptorForType().getFullName();
+
+    // lets serialize the message
+    int capacity = requestID.getId().length + data.length + messageType.getBytes().length + 8;
+    ByteBuffer buffer = ByteBuffer.allocate(capacity);
+    // we send message id, worker id and data
+    buffer.put(requestID.getId());
+    // pack the name of the message
+    ByteUtils.packString(messageType, buffer);
+    // pack the worker id
+    buffer.putInt(serverID);
+    // pack data
+    buffer.put(data);
+
+    return server.send(channel, buffer, capacity, 0);
+  }
+
   private class Handler implements ChannelHandler {
     @Override
     public void onError(SocketChannel channel) {
-      socketChannels.remove(channel);
+      workerChannels.remove(channel);
       connectHandler.onError(channel);
 
       loop.removeAllInterest(channel);
@@ -169,33 +235,36 @@ public class RRServer {
 
     @Override
     public void onConnect(SocketChannel channel, StatusCode status) {
-      socketChannels.add(channel);
+//      socketChannels.add(channel);
       connectHandler.onConnect(channel, status);
     }
 
     @Override
     public void onClose(SocketChannel channel) {
-      socketChannels.remove(channel);
+      workerChannels.remove(channel);
       connectHandler.onClose(channel);
     }
 
     @Override
     public void onReceiveComplete(SocketChannel channel, TCPMessage readRequest) {
-      // read the request id and worker id
-      // read the id and message
+      // read headers amd the message
       ByteBuffer data = readRequest.getByteBuffer();
-      byte[] id = new byte[RequestID.ID_SIZE];
-      data.get(id);
+
+      // read requestID
+      byte[] requestIDBytes = new byte[RequestID.ID_SIZE];
+      data.get(requestIDBytes);
+      RequestID requestID = RequestID.fromBytes(requestIDBytes);
 
       // unpack the string
       String messageType = ByteUtils.unPackString(data);
 
       // now get the worker id
-      int clientId = data.getInt();
+      int senderID = data.getInt();
 
-      RequestID requestID = RequestID.fromBytes(id);
+      // it can be a worker that will get its id from the job master
+      senderID = addToChannelListAssignIdIfRequired(channel, senderID);
+
       Message.Builder builder = messageBuilders.get(messageType);
-
       if (builder == null) {
         throw new RuntimeException("Received response without a registered response");
       }
@@ -204,23 +273,23 @@ public class RRServer {
         builder.clear();
 
         // size of the header
-        int headerLength = 8 + id.length + messageType.getBytes().length;
+        int headerLength = 8 + requestIDBytes.length + messageType.getBytes().length;
         int dataLength = readRequest.getLength() - headerLength;
 
+        // reconstruct protocol buffer message
         byte[] d = new byte[dataLength];
         data.get(d);
-
         builder.mergeFrom(d);
         Message m = builder.build();
 
         if (channel == null) {
           LOG.log(Level.SEVERE, "Chanel on receive is NULL");
         }
-        LOG.log(Level.FINEST, String.format("Adding channel %s", new String(id)));
+        LOG.log(Level.FINEST, String.format("Adding channel %s", new String(requestIDBytes)));
         requestChannels.put(requestID, channel);
 
         MessageHandler handler = requestHandlers.get(messageType);
-        handler.onMessage(requestID, clientId, m);
+        handler.onMessage(requestID, senderID, m);
       } catch (InvalidProtocolBufferException e) {
         LOG.log(Level.SEVERE, "Failed to build a message", e);
       }
@@ -230,5 +299,36 @@ public class RRServer {
     public void onSendComplete(SocketChannel channel, TCPMessage writeRequest) {
 
     }
+  }
+
+  /**
+   * add to channel list if this is the first message from this worker
+   * assign a worker ID if the job master assigns IDs
+   */
+  private int addToChannelListAssignIdIfRequired(SocketChannel channel, int senderID) {
+    // if the channel already exist, do nothing
+    // it means that the channel for this worker already added
+    if (workerChannels.containsKey(channel)) {
+      return senderID;
+    }
+
+    // if it is the submitting client
+    // set it, no need to check whether it is already set
+    // since it does not harm setting again
+    if (senderID == CLIENT_ID) {
+      clientChannel = channel;
+      return senderID;
+    }
+
+    // assign a new worker ID to this new worker
+    if (jobMasterAssignsWorkerID) {
+      int newWorkerID = workerChannels.size();
+      workerChannels.put(channel, newWorkerID);
+      return newWorkerID;
+    } else {
+      workerChannels.put(channel, senderID);
+      return senderID;
+    }
+
   }
 }
